requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
av==11.0.0
bitarray==3.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Cython==3.0.12
datasets==3.4.1
decorator==5.2.1
dill==0.3.8
distro==1.9.0
EfficientWord-Net==1.0.2
fairseq @ git+https://github.com/liyaodev/fairseq.git@b963eac7a04c539ad59fb1e23277f2ff7ee29e74
faiss-cpu==1.7.3
fastapi==0.115.11
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
geographiclib==2.0
gTTS==2.5.4
h11==0.14.0
huggingface-hub==0.29.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
invoke==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
loguru==0.7.3
lxml==5.3.1
MarkupSafe==3.0.2
mashumaro==3.15
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
numba==0.59.0rc1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.21.0
openai-whisper==20240930
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pastel==0.2.1
platformdirs==4.3.7
poethepoet==0.24.4
pooch==1.8.2
portalocker==3.1.1
praat-parselmouth==0.4.5
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
pyarrow==19.0.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2025.1
pyworld==0.3.5
PyYAML==6.0.2
regex==2024.11.6
requests==2.26.0
resampy==0.4.3
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
tflite-runtime==2.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchcrepe==0.0.22
tqdm==4.67.1
transformers==4.50.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.0
xxhash==3.5.0
yarl==1.18.3
typer==0.15.2